Fixed Grid i/j issue
It’s a bit awkward, but because printing has to be done row by row,
internally the x coordinate is the last one. This however is made
completely transparent to clients of the class, so using Grid you can
assume regular (x, y) coordinates. Square is a little redesigned for
the fact it can only have one impassable entity. Also TestSquare added
as it helped in debugging.